webui: setup subdirectories for the different components
The more components (UI elements) we implement the more obvious is the
need for directory organization. Start now with one level nesting,
following the modules directory naming.
